Fix #27: "regular" plots not shown if first row contains no result.
If a non-scatter, non-quantile plot should be shown in the HTML table,
and "correct results only" is selected, and the first column
of the first row does not contain a correct result,
the plot was empty because JQPlot could not manage a NaN value at that point.
This fixes the problem by inserting a different fake value.