Change how results of tools are analyzed.
There is now a constant result.RESULT_ERROR for an unspecified error.
Tool wrappers can return this to distinguish properly between UNKNOWN
and ERROR.

If the tool wrapper returns UNKNOWN or ERROR,
BenchExec itself now tries to provide a little bit more information,
e.g., whether the tool segfaulted, was killed, etc.

If there was a violation of the memory limit,
any result of the tool is now discard,
previously this was only done for timeouts.
However, in all such cases we now report the original result
as additional information like in "timeout (true)"
to allow detecting and debugging such cases
(typically these reflect bugs in the tool).