Create fresh temporary directories for every run and specify them in the $HOME and $TMPDIR variables to improve isolation of runs. For example, runs should not be influenced by existing files in $HOME (this would make it difficult to reproduce the results on other systems), and multiple runs creating files with the same names in shared home or temp directories would also influence each other.
This has only an effect if the tool uses these variables (as it should),
and does not hard-code "/tmp" or asks the OS for the home directory
of the user (which would return the original home directory).

Note that the Java Virtual Machine does not honor $TMPDIR and
hardcodes "/tmp". Developers of Java tools should read $TMPDIR in their
wrapper script and set the system property "java.io.tmpdir" with its value.

If really necessary, this behavior can be overridden in the tool wrapper.