Change behavior of "benchexec --timelimit" if "hardtimelimit" was specified in benchmark definition file.
Previously, the command-line parameter overwrote only the "timelimit"
value", not the "hardtimelimit" value if one was given.
Now if "--timelimit" is specified, this overwrites all time limits
(such that time limit and hard time limit are the same afterwards).
This should match user expectations of this parameter more closely.

Note that in general the command-line parameters are intended for
temporary adjustments and tests, for "real" benchmarks the parameters
should be specified in the XML file, which should be archived.