Add the column type to each RunSetResult's column once we have collected the set's data.
Current column types are 'text', 'measure' and 'count', 3 enums of type ColumnType.
If all rows of a column contain integer values, the column's type is 'count'.
If all rows of a column contain numbers and at least one value is a decimal, the column's type is 'measure'.
If at least one row of a column contains a value other than a number, the column's type is 'text'.

For columns of the type 'measure', the maximum amount of digits after the decimal point are stored, in regards to the number of significant digits.