Compute column types for statistics and use these for statistics formating.
Compute separate column types for statistics and do not use the column's already
computed type since this might result in wrong formatting if the max. number of
decimal digits differs.

Set an additional class based on the column type in the statistics section of html output.
Measure and count classes are aligned to the right for formatting, other classes are centered.