Different behaviour for explicit and implicit units of column values.
If an explicit unit is defined for a column in the table definition and a
different unit is found in one of its cells, raise an error.

If no explicit unit is defined for a column and two different units are found
in its cells, handle the column values like text.