Compress result XML files in benchexec, unless --no-compress-results is given.
This also optimizes XML writing somewhat, instead of creating a string
in memory, the XML is now written (and compressed) on the fly.

The temporary XML files that get created during benchmarking as
intermediate results won't be compressed even after this change,
and now they are also not longer pretty-printed to make writing them
more efficient.