Adjust format options for csv output.
If it is specified that a value should be aligned with csv as format target,
it is obviously intended for humans.
Because of this cut leading 0's for decimal values (0.1 -> .1)
and do not add html-whitespaces, but regular spaces.

A remaining problem is a space for the decimal dot for integers in a column
were decimal values exist.