Format very small/big float values correctly.
Convert the exponent representation of very small
and big float values to the decimal representation
after scaling.
Change mechanism for rounding to significant numbers to
always produce floats in decimal representation, only.

Also introduce a new format_target for tooltip values
to consider the special case of statistics values in tooltips
that are no stochastic values.

Since the column type for statistics columns
only considers the statistics values that will actually be
displayed in the column (i.e. 'sum'), this new type does not use the
column type's information for formatting values in the tooltips.
E.g. for a statistics column with only one cell, with sum = 10
and minimum value 3.33, the max number of decimal digits
will be computed using 10 only (which results in 0 decimal digits),
which later on would result in a wrong formatting for value 3.33
in the tooltip display.
The new format_target 'tooltip' works around this.