Table-generator: Add possibility to diff over arbitrary columns
In its current state, the table generator is able to create a table with
the differences between run sets.  This currently only considers the
"status" column.

This commit extends the functionality in order to use one (or more)
arbitrary columns that are defined in the table definition XML file for
the comparison for the diff table.

To activate the feature, you need version 1.7 of the table DTD and add
the attribute "relevantForDiff" to the column, for example:

    <column title="col1" relevantForDiff="true">Matching String</column>

If the attribute is not specified at any column, the status column will
be used to calculate the differences.  Otherwise only the columns with
the attribute present will be used.