Use synchronous signal handling in container mode.
Inspired by https://github.com/Yelp/dumb-init/pull/72.
In container mode, we need to have a process that acts as PID 1
and forwards all signals to the actual benchmarked process
(while at the same time waiting for this process to finish).
We can do this with sigwait instead of asynchronous signal handlers,
which is less and complex and avoids some signal handling problems.