Fix reading column values from zipped log files if relative paths are used.
This is a replacement for the fix in #141 that does not break other use cases.

The tests did not catch this bug because they only used absolute paths,
so we change them to relative paths (which typically make more problems
than absolute paths and are more common in real world usage).