Remove some broken preprocessing of files from tool-info modules.
Some tool-info modules created a temporary file with a modified input
file and gave this file to the tool.
They then stored the file name in an attribute of `self` and deleted it
in `determine_result`.
This is broken, because there can be further `cmdline` calls before the
call to `determine_result`, meaning that temp files won't be cleaned up.
On the other hand, calls to `determine_result` could delete temp files
that are still needed for other runs.

If a tool needs preprocessing, this should be done as part of the run in
a wrapper script. Only this approach leads to the resource usage for
preprocessing being measured and temp files being correctly removed.