Revert "Remove empty directory from upload"
This reverts commit 56b0ff7a1f5d78c4eddc75ea5f967663434de7cb.

The directory should not be empty. It was caused by some trick of the Archive
Manager that transformed the symbolic link (instrumentations ->
share/llvm-instrumentation) to a regular directory and left share
directory empty. The last archives has been archived with usual tar
command which leaves share non-empty and keeps the symbolic link.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>