Table generator: Reuse existing column types and use no initial one.
If our initial column type is 'count', as it was before,
and a column is empty, its type will be 'count', even
though we want it to be 'text' then.

If we use 'text' as initial column type, our code will be prone
to errors since then it must be possible to override a
text column type with a numeric column type - but we want
'text' to be a final type once it is decided that no other type
fits.