Replace dummy tools "true", "false", and "rand" by a new dummy tool "dummy" that is more general.
This new dummy tool can be made to output any result,
and even choose randomly between several choices,
so it can simulate any of the old dummy tools,
but is more flexible, which can be useful when debugging result
handling.

It relies on the tool "shuf", which is part of coreutils and should be
installed on all systems.