Output results for each block unconditionally
When only one block of benchmarks is executed,
the only output is the overall results
(with no "block" attribute in it).
However, when the results are processed automatically,
it is convenient to have an output for each block
no matter how many blocks we have.
This patch removes the special condition for single blocks
so that the output files have always the same structure.