Support "properties" dict in task template files.
This dict lists all properties the current task is expected to have.
If benchexec is given a property file,
and this property file is not listed in the task template,
that task template is ignored.

The properties dict also contains the expected result for each property,
but these are not used yet.