If a run has no SV-COMP property, set score to None instead of 0 in table-generator
This has only an effect if the column "score" is explicitly added to the
table in a table-definition file.
In this case, the column entries will now be "-" instead of "0"
for runs without SV-COMP properties,
and such runs will not be used for the statistics computation
of the score column (previously they were counted with value 0).
The same change occurs if there is a run with an SV-COMP property
but category "missing".

This fixes #404 and replaces #405.