Put temp directories for container on tmpfs (RAM disk, fixes #286)
This commit changes how the container mode works by default,
but this is ok because container mode is still in beta anyway.
If no RAM disk should be used, it can be disabled with --no-tmpfs.

We take care to force the kernel to put the tmpfs content on the same
NUMA node as the other memory of the process.
We also set the memory limit of the process as max size of the tmpfs.
The process will never be able to write that much onto the tmpfs
(because the memory limit covers tmpfs content + process RAM),
but any smaller limit might be too small, and any larger limit is useless.