Restructure the termination of a run
Previously, the child process (PID 1 in the container)
terminated and afterwards the parent did stuff like cleanup
and copying of result files.
Now the child waits until the parent signals that these tasks are
finished, and only then terminates.

This is necessary for the next changes which will add a tmpfs
for containers, and this tmpfs goes away if the child terminates,
but we need the tmpfs for copying result files from it.