Make the directory with the temporary files accessible in the container and copy result files from there instead of accessing the temp dir directly outside the container.
This is necessary for when we want to put the temp files into a tmpfs,
because this tmpfs will only be visible in the container.

Of course we do not want the tool to access these temp files directly,
so while we keep the temp dir accessible in the container,
we hide it below a mount point.
This means that the tool can only access it if --root is used
(otherwise it has no permissions to unmount the hiding mount),
and even then it is unlikely to cause problems
(the tool would need to unmount this on purpose).

Another slight disadvantage for this is that runexec
can no longer drop all privileges before the run is started,
it needs to keep the privilege to unmount the hiding mount
after the run is finished.