Fix #164: append crash dump information to tool output even in container mode
Some tool produce a file like hs_err_pid*.txt if they segfault,
and BenchExec detects this and appends the file content to the log file
of the run, because this is useful for the user.
So far, this did not work in container mode because the code that did
this was executed too late, when the container (and the files produced
within it) were already removed.

Now we do this step much earlier, almost immediately after the tool
exits and before the container is removed.