Add tool-info module for Frama-C
Frama-C allows to specify batches of commands
that are executed iteratively.
These batches are separated through a group of
`-then*` parameters, and each batch re-uses
the results of the previous batch.

Because of this, the original input file is usually
given to the first batch, and this tool-info
module does this, as well.