In container mode, mark the intermediate process as "non-dumpable"
The intermediate process is part of BenchExec,
but lives inside the container and this thus accessible
to the benchmarked process, which could for example use the process
to attempt escaping the container via one of the magic symlinks in /proc.

Marking the process as non-dumpable has the side effect
of forbidding most of the accesses from other processes.
This is best practice among container runtimes.