Make container.setup_user_mapping more flexible by adding parameters
This also has the side effect that os.getuid() and os.getgid()
are now no longer called by setup_user_mapping.
Instead they are called when the module is imported.
This can make a different if the current process changes its UID/GID in
between. In this case, use the new parameters to explicitly override
them.