Execute all methods of the tool-info module in container (fix #384)
In the past, there have been problems because the code in tool-info
modules sometimes changed files on disk (either directly, or indirectly
when it executed the tool itself for determining its version),
and this harms reproducibility.

Now the code of the tool-info module and all commands executed from it
are transparently executed in a container, just like the actual runs.
However, cgroups are not used so no resource limits are applied.

This is only done if --container is set (default in BenchExec 2.0),
and all options that modify the container setup (e.g., --network-access)
also apply to the tool-info module's container,
with the exception of --no-tmpfs (this container always uses a tmpfs).

This feature is also available in the test_tool_info utility:
python3 -m benchexec.test_tool_info --container <TOOL>