Remove auto-build from CPAchecker's tool-info module
Previously, the tool-info module of CPAchecker called CPAchecker's build
system if it detected that CPAchecker was run from a source checkout,
in order to ensure that an uptodate build was executed.

If the tool-info module is run within a container, this does not work
anymore (the freshly built files won't be visible outside the
container).
So we replace the automatic build with a check that looks for
modifications of the source and aborts if the sources are newer than the
binaries. This is actually just a heuristic and won't detect partially
successful builds, but it should be good enough in practice.

Because the check is only executed if both sources and binaries exist in
the project directory, it can be expected to not create problems if at
some point in the future the project structure of CPAchecker changes.

We also remove an existing warning if cpachecker.jar is missing.
Previously this warning would never be shown (because the automatic build
created the JAR file), and the JAR file is only necessary in specific
circumstances (VerifierCloud benchmarks) anyway, in which the benchmarks
will just fail otherwise (no possibility of hidden wrong results).