Even if lxcfs is used, give container acess to a fully visible proc instance
In theory, this breaks encapsulation a little bit:
Even though lxcfs provides a fake uptime, with this second visible proc
instance a process could in theory read the host uptime. However, the
information that is hidden by lxcfs is not crucial (such as the uptime),
and we hide the second proc instance such that normal processes will never
use it.

The advantage of this is that nested containers now work if lxcfs is
installed and used by BenchExec. Previously this required
--keep-system-config for the outer container to disable the use of
lxcfs. Now the following command works for me:
containerexec -- containerexec --hidden-dir /sys/kernel/debug /bin/bash