Clarify that containerexec --root allows access to hidden directories
Proof of concept:
containerexec --root -- /bin/sh -c "umount -l /run; find /run"

This parameter is only present in containerexec, and not in
benchexec/runexec, so it does not deserve a more prominent warning.
However, we still want to keep this parameter because for debugging it
is quite useful to be able to enter a container and fiddle with its
mount points.