Refactor argument passing of ContainerizedTool
Instead of passing *args and **kwargs around,
we collect those once in a list and a dict, pass these two around,
and unpack them only at the end.

This avoids possible name clashes in the future between argument names
of the tool-info module and the proxy functions in ContainerizedTool.

Also add some documentation.