Do not use ulimit for limiting CPU time
Previously, this was done as a fallback for the cgroup-based time limit.
If cgroups were not available, ulimit would at least provide some limit
(though unreliable), and if cgroups were available, ulimit would trigger
faster because it is not sampling-based.

We do not want ulimit without cgroups, because this is imprecise,
so we now just forbid time limits if cgroups are not available.
We also do not want ulimit in addition to cgroups.
While this would not be wrong, if ulimit kills our process we do not
know that the time limit was hit, so the terminationreason was wrong.