Fix bug for killing processes without container mode
If the freezer cgroup is available, we want to use it as protection
against fork bombs.
So far it was random which cgroup hierarchy was used first when killing
processes, so if the cgroups were mounted in separate hierarchies,
it could happen that we try to kill all processes using another
hierarchy first (without the fork-bomb protection of freezer) and only
afterwards make use of freezer.

Now we make sure that if freezer is available, we always use it first.