Remove support for tables with multiple status columns per run set
This was never documented and only used in SV-COMP in 2016 and 2017,
where there was one status column with the summarized result
and other status columns with the witness-validation result.
But even SV-COMP moved to a better way of representing these results,
and this feature has the unexpected side effect that all columns
with "status" appearing in the column title would be treated
as one of these additional status column.
So lets remove it. Any such column will be treated as plain text.

Reverts 962207f48d3fd4b7fcde417c1ce4e94f108cfb01
and acc209299ae58a23dbf0045c344a90f361e161f5