Refactor handling of Ctrl+C (SIGINT) in benchexec
Instead of using the default Python signal handler and checking for
KeyboardInterrupt, we can simply use the same mechanism that is used
for stopping BenchExec in other situations.

The difference is that with the new way of handling Ctrl+C,
it will just cancel all runs and otherwise continue executing normally.
The old way could interrupt benchexec in other situations
like writing output files.