Fix Ctrl+C in container mode
Previously, Ctrl+C killed the tool but only after a second Ctrl+C
BenchExec would terminate itself properly in container mode.

The reason for this was that apparently the Python interpreter
does a fork when we kill processes via the freezer cgroup
in the code added by 98d51abf.

The new child process inherits a copy of the file descriptors,
including those marked as noninheritable (not sure why).
One of these file descriptors is the pipe used for communication
between parent and our child process after the run finished.
The parent closed the pipe and the child did an os.read()
that was expected to terminate because no other end of the pipe was
still open.
This failed because the unexpected Python fork kept the pipe open.
Now we just communicate by sending something over the pipe.