Add documentation about not supporting KeyboardInterrupt in RunExecutor
KeyboardInterrupt can be raised at any point in time,
which makes it very hard to properly handle it in code
that needs to do a lot of cleanup (such as killing processes,
removing temp files etc. like we do),
especially given that repeated presses of Ctrl+C will again raise
KeyboardInterrupt in our cleanup code.

Thus we never supported it and require callers to either ensure
that RunExecutor is not used on the main thread (benchexec does this)
or that a signal handler for SIGTERM is installed that calls
RunExecutor.stop() (runexec does this).
Now we document this.