Fix problem in container mode when Python lib dir is not visible
This is triggered for example when using a virtual env in the home
directory and using "--hidden-dir /home", such as in some of our tests.

This import was removed in 500909534d672f405e4622ec3b807f104e9424f7.
We re-add it it container.py because the code that triggers lazy loading
of the resource module is in wait_for_child_and_forward_signals.