Allow ${inputfile_*} variables only if task-definition files are not used
This affects benchexec (in benchmark-definition files)
and table-generator (in table-defintiion files).

Previously, these variables would always be defined,
but contain the name of the task-definition file
instead of some input file, which can be confusing.

As before, the variables ${taskdef_*} can be used
whenever a task-definition file is used.

This fixes #390 as discussed in
https://github.com/sosy-lab/benchexec/issues/390#issuecomment-501127078

This is a backwards-incompatible change,
but the use of undefined variables already causes a warning.