Workaround for deadlock of benchexec in container mode with parallel benchmarks
Fixes #435.

This would occur for example if executing 100 parallel runs.
This leads to a large number of threads and high thread contention,
and forking the interpreter during thread contention
(when gil_drop_request is set) deadlocks when re-entering the Python world
in the child process. More information is in the issue.