Add a real fix for #435, at least for Linux x86_64
The best way for fixing #435 is to use a callback function for clone
that does not contain Python code. We do not want to add a C module
to BenchExec because of the trouble with building and distributing it
(for example, the C code would depend on the Python version).
So we dynamically generate machine code on the fly that does the same
as our Python function but without the deadlock problems.
This may sound a little bit crazy, but it is actually quite easy and the
corresponding assembler code is relatively simple (just 9 instructions)
and well understandable with some basic assembler knowledge.
For now we only provide an implementation for Linux x86_64 and use the
existing workaround for other architectures.
Of course this could be extended to other architectures in the future.