Fix #447 of CPAchecker's tool-info module
When CPAchecker reported a property as violated,
it would always return "false(unreach-call)" except for a list of
hard-coded properties, even if some completely different property was
violated.

Now it returns "false(something)" if CPAchecker outputs
"Property violation(something: ...)" and "false" otherwise.

In practice this has the effect that for current CPAchecker versions the
output "false(unreach-call)" will be replaced with "false",
but BenchExec accepts both results anyway.
If we want CPAchecker's tables to contain "false(unreach-call)" as
before, we should change its internal specification to add
"unreach-call" to the output.