Add infrastructure for seccomp filters in containers
This allows us to filter syscalls.
It has a small performance impact, but it is deterministic and probably
only small (other container runtimes like Docker also do this).

For now, no syscalls are explicitly filtered,
but the default filter already forbids syscalls from binaries that have
a different architecture than the kernel's native architecture.
We whitelist 32bit binaries on 64bit kernels for x86 and ARM, though,
which are probably the most common cases of non-native binaries.