Fix critical bug of container for tool-info module (since BenchExec 2.1)
If an error occurred during creation of the container,
the tool-info module would silently be used without containerization
instead of showing the error to the user.

When an exception occurs during creation of the container,
we would previously catch and return the exception.
However, 30665e88 led to some exceptions being silently swallowed.
Now we just raise exceptions, which the multiprocessing module handles fine.

As a side effect, the documentation of the tool-info module
was missing when executing benchexec.test_tool_info.