Attempt to add more debugging info if loading of tool-info module fails
Fixes #459 as far as I know that this is possible.

The information printed to the debug log is not necessarily correct.
The Python module-loading system is quite complex and there are things
like path finders, meta path finders, and import hooks that can all
influence how packages are loaded.
Because arbitrary code can be executed, there is no way to print exact
information where packages are searched.
However, in a standard scenario Python looks in the paths contained in
sys.paths, and we can print these. This should help most users.