Improve passing data from table-generator to react-table
If formatted and original value are the same, we do not duplicate them
anymore. This helps to make the file size smaller.

We also do not pass most column values as formatted value anymore
(which is interpreted as HTML code), because this could cause
misinterpretations of characters and HTML injection attacks.
Only HTML code that is produced by table-generator's column formatting
(and that is safe) is passed as formatted value.

Also rename "original" to "raw" and "formatted" to "html".