Add local scrollbar to "Environment" table of react-table if necessary
Previously, this table would overflow and the browser would show a
scrollbar for the whole window. This would mean that one would have to
scroll the environment table (which is at the top) with a scroll bar at
the bottom of the window, and the statistics table (which is at the
bottom) with its own scroll bar.

Now both tables have their own scroll bar just below of them.

Also rename both tables to make their name more clear
("Benchmark Setup" instead of "Environment",
"Statistics" instead of "Summary" because the latter is already the name
of the tab).

This fixes #491.