Simplify JSON passed to react-table
Instead of passing both "short_filename" and the full task name
(as part of the "id" array), we just pass what is needed.
Furthermore, we add the "href" attribute for rows only if it makes
sense, and can therefore remove the "has_sourcefile" attribute,
and we pass only those parts of the "id" array that are relevant.