Highlight focused cells in the react table with a red dotted border
The goal is to have a visual indication which cell was clicked
after closing the overlay with the link content
(the focus is automatically restored by react-modal).
Firefox does this anyway with the typical dotted grey focus marker,
but Chromium doesn't...
So we set a specific focus marker, and while we are at it,
we can make it better visible.

It is now similar to the behavior of the old table
(although there, the marker was not coupled to focus,
but would stay until a different cell was clicked).

Fixes #494