Fix crash in RunExecutor if tool produces invalid UTF8 on stdout
RunExecutor parses the output of the tool to find out whether there is
some crash dump information in another file.
So far it assumed that the tool output is valid text in the encoding of
the system, but some tools apparently dump binary stuff to stdout.
Now we treat the tool output consistently as a sequence of bytes.