Fix sorting and normalization of data points in quantile plot
If a graph does not have a data point for every row in table
(for example because of missing results, or a "correct only" plot),
the data points were put on adjacent x values in the "direct" plot
instead of at their proper x coordinates.
Furthermore, sometimes values like NaN were produced and shown,
which additionally broke the sorting of the quantile plot.

Now we properly normalize all invalid values and hide them,
and compute proper x values even if values are missing.

Fixes #500