Replace JSZip in react-table with zip.js
We need a ZIP library for extracting logs from ZIP archives,
and we do not want to download the whole ZIP archive for that.
zip.js supports partial extraction of some files and downloads
only the necessary parts.
There seems to be no other library than zip.js that supports this
feature.
Fixes #501

The disadvantage of zip.js is that it is not available as a npm package
and that it is not a proper module, so we have to commit it directly in
this repository and jump through some hoops to be able to use it,
but it should work fine now that this is done.

A small advantage is that zip.js has no dependencies
and now the final bundle of JS is about 100 KB smaller (uncompressed).

The code that loads the logs from ZIP archives is copied from the old
template.html and should be rewritten in style of modern JavaScript,
but we can do that later.
This code also has some extra fallback for Chrome that fixes #499.