Disable a webpack option that leads to vendors.min.js changing more often
occurrenceOrder is advertised as reducing the resulting file size,
but for us it has no positive effect on file size.
Furthermore, it increases the chance that vendors.min.js changes
if we change our own code, and we want to change that file as rarely as
possible to avoid increasing the size of the git history more than
necessarily.

Part of #482