Refactoring of some components in react-table and more consistent CSS
Several things were duplicated, e.g., the column headers,
and are now moved into separate components and reused.
Some CSS was inconsistent between the statistics table and the main
table, and the header of status columns mistakenly had the "btn" CSS
class.
Also move the stateless function getRunSets from Overview (where we keep
the state) to utils.js, where other such functions are already.

The change to vendors.min.js is only because we are using destructuring
assignment with a "rest" parameter ({foo, ...rest}) and a new helper
function is needed for that.