Improve and extend metaval tool-info module
There was a bug in the delegation of determine_result
to the wrapped tool. Some tools like Symbiotic determine the version
again in this step and need access to the binary again. If we do not
change into the wrapped tool's folder, the binary will not be found.

This change also adds some configuration options. The first one is
needed for running older versions of wrapped tools which need Java 8
and therefore a custom entry in the PATH variable. The second one
will enable the tool user to ensure that only a certain type of
witness is checked.