Distinguish missing results in tables produced by table-generator
There are two big classes of missing results.
The first is if we combine several results that were produced for
different benchmark sets. In this case we now assign the category
"empty" and continue to show an empty cell as status in the HTML table.

The second is if there is a respective <run> tag in the result file,
but it does not contain any results. This happens for example if
benchexec is interrupted with Ctrl+C. In this case we now assign the
category "aborted" and show a dash as status in the HTML table.

We also allow to filter for these two cases in HTML tables
using the existing status filter.

In CSV tables, the status value is still empty.
Users who want to distinguish these cases in CSV tables
should add the column "category", which now has the two new values
described above.