Task filters for benchexec based on the expected verdict
Implementation of the most important part of #529.
The syntax is <propertyfile expectedverdict="...">,
where the value can be either "true", "false", or "unknown".
Filtering on anything else is not implement.
The filters also work only for tasks that are defined in
task-definition files, not for the deprecated kind of tasks
where the expected verdict is encoded in the file name.

No tests exist yet.