Add a tag <description> to result XML files with user-specified content.
Users can provide a text file with a description of the benchmark
(e.g., who executed it, which version of the benchmark set was used)
by using "benchexec --description-file <text-file> ...",
and benchexec will copy the content of that file into the produced
XML files with the results.
This allows to properly document important facts about the benchmark
results even if BenchExec itself does not know about them.

In the future, table-generator will take the information from the
<description> tag and present it in the HTML tables.

First part of #524.

Replaces f3a27520 from #526.