Add starttime and endtime attributes to result XML files
The new fields replace the date attribute (which is still added for
backwards compatibility).
The new fields are in ISO format, have subsecond precision,
and are specific per run set
(when executing several run sets out of a single benchmark definition,
the date attribute has the same value for all result files,
the new attributes show precisely when benchmarking of each run set
started and ended).
For tests, where we need deterministic output,
we do not output endtime if "benchexec --startTime" is used.
Of course, the value of that parameter will also override the new
starttime attribute, just as it does with the date attribute.

Fixes #221.