Add script to transform old-style tasks to yml-task definition format
Takes C files or .set files that contain glob-list of C files.
Renames C files with '_true-PROPERTY' and "_false-PROPERTY"
verdicts in filenames to not contain these verdicts anymore.
If two files handled in the same call to this script have a same new
name,
the script adds a suffix '-1', '-2', etc. to the new names to avoid
colissions.

Next to each renamed task, a yml-file with the same name as the C file
is created. The yml file links to the renamed file and lists all
properties that were previously encoded in the file name.
This yml-file is the new way to define tasks,
cf. https://github.com/sosy-lab/benchexec/blob/master/doc/benchexec.md#task-definition-files