Do not overwrite result UNKNOWN with ERROR (1) in benchexec
Fixes #538

If the tool-info module returns a result like "ERROR" or "UNKNOWN",
benchexec enriches this if possible, e.g., by adding the return code.
However, some tools use a return code to mark UNKNOWN results,
and these would be shown as error.
So we no longer do this for the case of UNKNOWN.