Fix rudimentary support of benchexec for non-Linux systems
Actual benchmarking is not supported,
but there are extensions of benchexec that work on non-Linux systems
and we try to accomodate them by importing modules that require Linux
only if necessary.
This logic was broken unfortunately, and is now fixed.
Furthermore, we explicitly require --no-container as argument in this
case.