Remove support for expected verdicts of old-style tasks from table-generator
Part of #439.
For tasks that are not defined with yaml files,
table-generator so far parses the file name to detect the expected
verdict. This commit removes this.
The result is that for tables with such tasks,
there are no numbers of true/false tasks,
and no statistics for correct/wrong results.