Refactor how statistics are prepared for HTML table in tablegenerator
So far the formatting and preparation for insertion as JSON
was spread over two parts of the code, and datastructures were created
by the first part just to be converted into something else in the second
part.
Now we unify this code in the htmltable module and do everything in one
step.