Fix crash of table-generator if variable LANG is empty
We used the system encoding for writing out the HTML file,
and if LANG was not set, the encoding is ASCII and our special space
characters lead to a crash.
But using the system encoding for the HTML file is wrong anyway,
because the HTML file declares itself to be UTF-8.