Show property in HTML table if it is identical for all tasks
In this case, we show it in a row on the summary tab.
In all other cases, it is already shown in the first column of the table tab.

There was already code intended for showing the property on the summary
tab, but it did not work (and probably never did).

Fixes #395