Store Property and ExpectedResult instances in TaskId
Storing the Property instance instead of the string name
has the advantage that we do not need to parse the task-definition file
once again when creating Row instances.

Adding the ExpectedResult instance also makes it part of the unique
identifier of a task, i.e., task that differ only in the expected
verdict would now be treated as distinct.
This does not occur currently because we read the expected verdict
from the task-definition file (which is already part of the id),
but it can happen in the future when we read expected verdicts
from result files, so it is good to be prepared for this.